refactor(dashboardeditor): general refactoring, comment out advanced features
