+++ src/components/CardEditor/CardEditor.test.jsx
@@ -48,10 +48,8 @@ describe('CardEditor', () => {
       title: `${defaultCard.title}z`,
     });
     actions.onChange.mockReset();
-    userEvent.click(
-      screen.getByRole('button', { name: `Card size ${defaultCard.size} Open menu` })
-    );
-    userEvent.click(screen.getByText('MEDIUMWIDE'));
+    userEvent.click(screen.getByRole('button', { name: `Size Small (4x1) Open menu` }));
+    userEvent.click(screen.getByText('Medium wide (8x2)'));
     expect(actions.onChange).toHaveBeenCalledWith({
       ...defaultCard,
       size: 'MEDIUMWIDE',
@@ -89,7 +87,7 @@ describe('CardEditor', () => {
     });
     expect(openEditorBtn).toBeTruthy();
     userEvent.click(openEditorBtn);
-    expect(screen.getByRole('dialog', { name: 'Edit JSON' })).toBeTruthy();
+    expect(screen.getByText('Edit card JSON configuration')).toBeTruthy();
     userEvent.click(screen.getByRole('button', { name: 'Close' }));
     userEvent.click(openEditorBtn);
     userEvent.click(screen.getByRole('button', { name: 'Cancel' }));
