+++ src/components/DashboardEditor/_dashboard-editor.scss
@@ -8,10 +8,12 @@
   }
   &--header {
     flex: 0;
-    padding: $layout-01 0 $layout-02;
+    padding: $spacing-05 0 $spacing-03;
     .#{$prefix}--grid {
       /* allow the header to stretch the full page width */
       max-width: none;
+      padding-left: $spacing-07;
+      padding-right: $spacing-05;
     }
     .header-top {
       height: $layout-02;
@@ -22,6 +24,11 @@
       height: $layout-03;
       display: flex;
       align-items: center;
+      .#{$prefix}--btn--sm {
+        margin-left: $spacing-05;
+        padding-left: $spacing-05;
+        padding-right: $spacing-05;
+      }
     }
     .header-left {
       display: flex;
